Refactor and clean up
Use null for no wallet state instead of undefined.
Move helper functions to a utils module. We will likely use these in the
backups page.
Oragnize functions in transactions component and use updateTransactions for
initializing transactions table.
